// src/components/ContainerBreakpoint/types.ts
export enum EBreakpoint {
  // Source: https://mui.com/customization/breakpoints/
  // Breakpoints can be changed by https://material-ui.com/customization/breakpoints/#custom-breakpoints
  XSMALL = "xs",    //     0px
  SMALL = "sm",     //   600px
  MEDIUM = "md",    //   900px
  LARGE = "lg",     //  1200px
  XLARGE = "xl",    //  1536px
}